Extract a type.
